% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiomics-extraction.R
\name{extract_radiomics}
\alias{extract_radiomics}
\title{PyRadiomics Feature Extraction}
\usage{
extract_radiomics(image, mask)
}
\arguments{
\item{image}{File path to input image.}

\item{mask}{File path to nodule mask.}
}
\description{
PyRadiomics Feature Extraction
}
