% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integralrad-package.R
\docType{package}
\name{integralrad-package}
\alias{integralrad}
\alias{integralrad-package}
\title{integralrad: Pulmonary Nodule Malignancy Prediction using INTEGRAL-Radiomics}
\description{
An implementation of the INTEGRAL-Radiomics pulmonary nodule malignancy prediction model. The package enables standardized preprocessing, radiomics feature extraction, and malignancy risk prediction for pulmonary nodules detected on computed tomography (CT) imaging. It is designed to support reproducible research for radiomics-based lung cancer risk assessment workflows. The INTEGRAL-Radiomics model is described in Warkentin et al. (2024) \doi{10.1136/thorax-2023-220226} (PMID: 38195644).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mattwarkentin/INTEGRAL-Radiomics}
  \item Report bugs at \url{https://github.com/mattwarkentin/INTEGRAL-Radiomics/issues}
}

}
\author{
\strong{Maintainer}: Matthew T. Warkentin \email{matthew.warkentin@ucalgary.ca} (\href{https://orcid.org/0000-0001-8730-3511}{ORCID}) [copyright holder]

}
\keyword{internal}
