% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict_integral_radiomics}
\alias{predict_integral_radiomics}
\title{INTEGRAL-Radiomics Nodule Malignancy Predictions}
\usage{
predict_integral_radiomics(input)
}
\arguments{
\item{input}{Path to a CSV. See Details.}
}
\value{
A \code{tibble::tibble} with the input columsn and one column added for
the probability of nodule malignancy (i.e., lung cancer).
}
\description{
Predict the probability of nodule malignancy using the INTEGRAL-Radiomics
model. The model uses the Python library \code{PyRadiomics} to perform feature
extraction before making model predictions.
}
\details{
The input for this function should be a path to aCSV with the following
columns (names must be exact matches):
\itemize{
\item \code{image}: Path to image (NRRD format)
\item \code{mask}: Path to nodule mask (NRRD format)
\item \code{age}: Age (years)
\item \code{sex}: Sex (0=Male, 1=Female)
\item \code{bmi}: Body mass index (kg/m^2)
\item \code{fhlc}: Family history of lung cancer (0=No, 1=Yes)
\item \code{copdemph}: History of COPD or emphysema (0=No, 1=Yes)
\item \code{formersmk}: Former smoker status (0=No, 1=Yes)
\item \code{duration}: Number of years smoked
\item \code{cigday}: Average number of cigarettes smoked per day
\item \code{quittime}: Years since quitting for former smokers (set to 0 for current
smokers)
}
}
