% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-radiomics.R
\name{extract_radiomics}
\alias{extract_radiomics}
\title{PyRadiomics Feature Extraction}
\usage{
extract_radiomics(image, mask)
}
\arguments{
\item{image}{File path to input image (NRRD format).}

\item{mask}{File path to nodule mask (NRRD format).}
}
\description{
This function performs the radiomics feature extraction using the
\href{https://pyradiomics.readthedocs.io/en/latest/}{\code{PyRadiomics}} Python
library. These 2,060 radiomic features are required by the
\strong{INTEGRAL-Radiomics} model. Configuration of the feature extraction is
automatically handled by the configuration file shipped with this package.
All Python dependencies are managed by \code{\link[reticulate:py_require]{reticulate::py_require()}}. This
function is exported so users can get the radiomic features, if desired.
However, most users will use \code{\link[=predict_integral_radiomics]{predict_integral_radiomics()}} directly which
calls this function internally.
}
\examples{
\dontrun{
extract_radiomics("image.nrrd", "mask.nrrd")
}

}
