% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict-integral-radiomics.R
\name{predict_integral_radiomics}
\alias{predict_integral_radiomics}
\title{INTEGRAL-Radiomics Nodule Malignancy Prediction}
\usage{
predict_integral_radiomics(input, quiet = FALSE)
}
\arguments{
\item{input}{Path to a CSV. See Details for more information on the format.}

\item{quiet}{Logical. Set to \code{TRUE} to suppress messages and progress bars.
Default is \code{FALSE}.}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble}} with the \code{input} columns and two columns added
for the probability of nodule malignancy (\code{pred_malignant}) and the
complementary probability of being a benign nodule (\code{pred_benign}).
}
\description{
Predict the probability of malignancy for a pulmonary nodule using the
\strong{INTEGRAL-Radiomics} model. This function uses the Python library
\href{https://pyradiomics.readthedocs.io/en/latest/}{\code{PyRadiomics}} to perform
feature extraction before making model predictions.
}
\details{
The \code{input} for this function should be a path to a CSV with the following
columns (column names must be exact matches):
\itemize{
\item \code{image}: Path to image (NRRD format)
\item \code{mask}: Path to nodule mask (NRRD format)
\item \code{age}: Age (years)
\item \code{sex}: Sex (0=Male, 1=Female)
\item \code{bmi}: Body mass index (kg/m^2)
\item \code{fhlc}: Family history of lung cancer (0=No, 1=Yes)
\item \code{copdemph}: History of COPD or emphysema (0=No, 1=Yes)
\item \code{formersmk}: Former smoker status (0=No, 1=Yes)
\item \code{duration}: Number of years smoked
\item \code{cigday}: Average number of cigarettes smoked per day
\item \code{quittime}: Years since quitting for former smokers (set to 0 for current
smokers)
}
}
